multiload: initial support for configuraiton
Add support for providing a configuration for the loader.  The configuration
file is a custom format composed of a set of constraints and a loader specific
for an architecture.  An example is:

  loader /usr/armv7eb-unknown-linux-gnueabihf/lib/ld-linuxhf.so.3 {
    arch arm
    subarch v7
    endian big
    flags hard-float
  }

The architecture specific validation is yet to be implemented, but, this is
sufficient to use a configuration file to dispatch the loader rather than
hardcoded paths.